Add argument checking to LongSequence.

+++ src/main/java/org/d2ab/util/Preconditions.java
@@ -7,32 +7,32 @@ public abstract class Preconditions {
 	Preconditions() {
 	}
 
-	public static void requireAtLeastZero(int value, String name) {
+	public static void requireAtLeastZero(long value, String name) {
 		requireAtLeast(0, value, name);
 	}
 
-	public static void requireAtLeastOne(int value, String name) {
+	public static void requireAtLeastOne(long value, String name) {
 		requireAtLeast(1, value, name);
 	}
 
-	public static void requireAtLeast(int threshold, int value, String name) {
+	public static void requireAtLeast(long threshold, long value, String name) {
 		if (value < threshold)
 			throw new IllegalArgumentException("Expected " + name + " to be >= " + threshold + ": " + value);
 	}
 
-	public static void requireBelow(int threshold, String thresholdName, int value, String name) {
+	public static void requireBelow(long threshold, String thresholdName, long value, String name) {
 		if (value >= threshold)
 			throw new IllegalArgumentException("Expected " + name + " to be < " +
 			                                   thresholdName + " (" + threshold + "): " + value);
 	}
 
-	public static void requireAtMost(int threshold, String thresholdName, int value, String name) {
+	public static void requireAtMost(long threshold, String thresholdName, long value, String name) {
 		if (value > threshold)
 			throw new IllegalArgumentException("Expected " + name + " to be <= " +
 			                                   thresholdName + " (" + threshold + "): " + value);
 	}
 
-	public static void requireNotEqual(int exception, int value, String name) {
+	public static void requireNotEqual(long exception, long value, String name) {
 		if (value == exception)
 			throw new IllegalArgumentException("Expected " + name + " to be != " + exception);
 	}
